extract signing_root module and adjust code to the change
