+++ ethereum-common/src/lib.rs
@@ -8,6 +8,7 @@ pub mod merkle;
 pub mod network;
 pub mod patricia_trie;
 pub mod rlp_node_codec;
+pub mod signing_root;
 pub mod utils;
 
 #[cfg(not(feature = "std"))]
@@ -41,3 +42,5 @@ pub const EPOCHS_PER_SYNC_COMMITTEE: u64 = 256;
 pub const SYNC_COMMITTEE_SIZE: usize = 512;
 /// According to Ethereum spec [v1.4.0](https://github.com/ethereum/consensus-specs/blob/v1.4.0/specs/altair/light-client/p2p-interface.md#configuration).
 pub const MAX_REQUEST_LIGHT_CLIENT_UPDATES: u8 = 128;
+/// According to Ethereum spec [v1.4.0](https://github.com/ethereum/consensus-specs/blob/v1.4.0/specs/altair/beacon-chain.md#domain-types)
+pub const DOMAIN_SYNC_COMMITTEE: [u8; 4] = [0x07, 0x00, 0x00, 0x00];
