+++ gear-programs/checkpoint-light-client/src/wasm/mod.rs
@@ -7,8 +7,9 @@ use io::{
         base_types::{Bitvector, FixedArray},
         merkle,
         network::Network,
+        signing_root,
         tree_hash::TreeHash,
-        utils as eth_utils, Hash256, SYNC_COMMITTEE_SIZE,
+        utils as eth_utils, Hash256, DOMAIN_SYNC_COMMITTEE, SYNC_COMMITTEE_SIZE,
     },
     meta,
     sync_update::Error as SyncCommitteeUpdateError,
